requirements.txt
Django==1.9.8
django-password-reset==0.9
djangorestframework==3.4.0
gunicorn==19.6.0
imutils==0.3.6
oauthlib==1.0.3
PyMySQL==0.7.1
pyparsing==2.1.0
python-dateutil==2.4.2
pytz==2015.7
queuelib==1.4.2
recordtype==1.1
requests==2.8.0
requests-oauthlib==0.6.0
Twisted==15.5.0
urlparse2==1.1.1
whitenoise==3.2.1
Pillow==2.7.0
psycopg2
dj-static
cloudinary